transite/requirements.txt
asgiref==3.3.1
cycler==0.10.0
dataclasses==0.8
Django==3.1.5
django-ckeditor==6.0.0
django-js-asset==1.2.2
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
typing-extensions==3.7.4.3
